requirements.txt
billboard.py
lyricsgenius
nltk
scikit-learn